Added modifiable player damage, hp gained from killing a mob, or piercing bullets (don't get destroyed on contact with enemy bullets, or swarm).

+++ PolygonPlatformer/Creature.cpp
@@ -4,11 +4,11 @@ SceneNode* Creature::currentPlayer = nullptr;
 
 
 
-void Creature::damage (unsigned amount) {
+void Creature::damage (float amount) {
     currentHP -= amount;
 }
 
-void Creature::heal (unsigned amount) {
+void Creature::heal (float amount) {
 	currentHP += amount;
 }
